Fix #22 Use importer callbacks to handle static path prefixes
Instead of import paths, use the importer callback so we can handle
static paths prefixes, and possibly also non-filesystem storage.

This is more complicated than anticipated, because we have to re-create
the fallback mechanisms for handling partials and extensions, and
because finders.find ignores prefixes.

There is probably room for improvement, especially around the source
maps, but this is a start.

Note that the test_raw_css_import test had to be changed, because
the libsass doesn't seem to use @import when callbacks are used.